move color mode to config
