+++ src/LogExpert/Controls/LogTabWindow/LogTabWindow.cs
@@ -254,6 +254,5 @@ namespace LogExpert.Controls.LogTabWindow
 
             #endregion
         }
-
     }
 }
