+++ src/LogExpert/Program.cs
@@ -98,7 +98,7 @@ namespace LogExpert
                 Settings settings = ConfigManager.Settings;
                 Mutex mutex = new Mutex(false, "Local\\LogExpertInstanceMutex" + pId, out var isCreated);
 
-                LogExpert.Interface.ColorMode.LoadColorMode();
+                LogExpert.Config.ColorMode.LoadColorMode();
 
                 if (isCreated)
                 {
